nfc_activity/pyproject.toml
[tool.poetry]
name = "nfc_activity"
version = "0.1.0"
description = "Access control and smart lock system"
authors = ["salieri256"]

[tool.poetry.dependencies]
python = "^3.10"
nfcpy = "^1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
